Extract AudioPlayer
