+++ app/src/main/java/com/legendai/musichelper/ui/MusicScreen.kt
@@ -1,14 +1,10 @@
 package com.legendai.musichelper.ui
 
 import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.PlayArrow
-import androidx.compose.material.icons.filled.Pause
 import androidx.compose.material.icons.filled.Settings
 import androidx.compose.material.icons.filled.List
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.platform.LocalContext
-import com.google.android.exoplayer2.ExoPlayer
-import com.google.android.exoplayer2.MediaItem
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.background
 import androidx.compose.foundation.lazy.LazyColumn
@@ -173,32 +169,6 @@ fun MusicScreen(
                     viewModel.mixAndExport(LocalContext.current, selected)
                 }) { Text("Mix & Export") }
             }
-        }
-    }
-}
-
-@Composable
-fun AudioPlayer(url: String, label: String) {
-    val context = LocalContext.current
-    var playing by remember { mutableStateOf(false) }
-    val player = remember(url) {
-        ExoPlayer.Builder(context).build().apply {
-            setMediaItem(MediaItem.fromUri(url))
-            prepare()
-        }
-    }
-    DisposableEffect(player) {
-        onDispose {
-            player.release()
-        }
-    }
-    Row(verticalAlignment = Alignment.CenterVertically) {
-        Text(text = label, modifier = Modifier.weight(1f))
-        IconButton(onClick = {
-            playing = !playing
-            if (playing) player.play() else player.pause()
-        }) {
-            Icon(if (playing) Icons.Default.Pause else Icons.Default.PlayArrow, contentDescription = null)
         }
     }
 }
